requirements.txt
google-api-python-client>=2.97.0
google-auth>=2.22.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.1.0
pandas>=2.1.0
numpy>=1.27.0
torch>=2.1.0
keyboard>=0.13.5
google-api-python-client 
google-auth 
google-auth-oauthlib 
PyQt5 
torch 
pandas 
pyttsx3 
transformers